requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
python-multipart
jinja2
reportlab
pydantic
aiofiles
